moving methods to APIHandler
